Outline workflow for importing & maintaining data
Problem:

Maintaining structured data is hard,
especially when we don't know the final format of the data
ahead of time.

We will have to deal with complications including:

- Data from different data sources in different formats
  (plain text, csv, etc)
- Data being updated after the fact in the original source format
- Changes to organization attributes
  - Addition
  - Removal
  - Rename
  - Restructure

Solution:

We will use semantic versioning for the data
and a series of scripts to help keep the data updated.

We will write automated tests to make sure
that data follows the version of the schema that it claims to be in.